Implement pagination (#21)
This contains several significant changes:
In the postgres backend, a created_at field is introduced to have a
stable ordering of assets.
The ledger pagination is pretty similar to what already exists in the
current chaincode.
We are constrained to the capabilities of the ledger, so this is not a
"real" pagination but rather a way to drain a sequence in multiple
calls. Each paginated response contains a token to get the next page.

Working on the pagination revealed an issue with the way contractapi
validates the inputs/outputs of each smart contract. That lead to a
broken invocation when any field is optional: protobuf empty fields
are omitted while contractapi expects all fields to be present.
We can bypass contractapi's limitations by forwarding json
un/marshalling to protojson for every asset. This is done at build time
by a custom code generation.