feat: implement error conversion in distributed mode (#45)
This will return proper gRPC error codes in distributed mode.
The logic is the same than in standalone mode: an interceptor attempts
the conversion. The only difference is on the conversion
implementation: where standalone has access to the error stack, we can
only rely on the message in distributed mode. That's why sentinel
values for orchestration errors have been assigned a unique
alphanumeric identifier.