fix: disable intermediary model
This commit adds a new end to end test to make sure intermediary
models can be disabled when the compute plan allows it.
It also fixes two related issues: the DeleteIntermediaryModels flag of
a new compute plan was ignored and there was a typo in the model
update SQL query.