fix: rework concurrency limiter strategy
This new implementation work at the handler level and should not
impact healthcheck or non-orchestration gRPC methods.
The whole strategy is to have a scheduler maintaining a FIFO queue of
pending requests. Each request wait for a token to be executed and
only the scheduler can assign such token, in a deterministic order.