feat(standalone): asynchronous dispatch (#211)
Since the dispatch of event occurs after transaction commit, even if
the events are not properly dispatched, we have no way to rollback the
changes. So the best we can do is to log and/or return an error.

Before this commit, the error was logged and we returned an error.
Now, there is only a log but we return before even actually
dispatching the events (since we can't do much anyway).