feat: do not cascade canceled status (#313)
This alter the behavior when a task is marked as failed or canceled.
When a task was marked as failed or canceled, we propagated a canceled
status to its children. This is not necessary to prevent their
execution since there is no way for them to switch to "TODO" status.
As such, no event will be ever dispatched, so the workers will never
process the children of failed tasks, no matter their status.
Getting rid of this propagation comes with a huge performance
increase: under some circumstances a cancellation had to walk the whole
task graph. Now it only impacts a single task.