feat(distributed): reuse gateway connection (#324)
This commit introduces a new gateway module with two components:
the Pool which handle multiple gateways and allows to reuse them
across multiple requests; and the Gateway which is mostly a wrapper
around the fabric's own gateway abstraction.

This allows to have a neat asynchronous API to query the chaincode:
on request arrival a Gateway is retrieved from the Pool (created if
needed) and passed to the chaincode Invocator (which hasn't changed
much). The Invocator requests a chaincode invocation from the Gateway
and wait until it receives an error or an output, whichever comes
first.