refactor: orchestration error handling (#360)
This commit switches from several sentinel values to an error wrapping
type. This change is prompted by the fact that we want to be able to
identify both the underlying error (triggering a retry on some
transactions) and the higher level orchestration error (to return an
appropriate response code).

Also removed unused ErrReferenceNotFound error.