chore: switch logger library for e2e tests
We used go-playground/log, which has low adoption and the maintainer
[explicitly mentions that he maintains it as a best
effort](https://github.com/go-playground/log/tree/v7.0.2#package-versioning).
Considering this, and the fact that the most up to date version
dropped support for a console handle, we should move towards a better
supported logging library.

This commit only affects e2e tests, but the goal is to switch to the
same library in application code.

I've tested both zap and zerolog, which are the two most prominent
contenders in the space of structured logging. Both have strong
communities and similar use. They both compete to be the fastest
logging library with the lowest memory allocations.

Comparing between the two: zerolog supports the singleton logger
pattern, while zap prefers the "pass the logger everywhere". (to be
fair, the singleton is also supported but discouraged in the doc.)

Other than the obvious API change, the main shift with zerolog is that
fields are typed. That's part of their recipe for performances since
most of the type reflection is not involved. In cases where we want to
log structures (mainly proto messages), the `.Interface()` method will
rely on reflection.

Zerolog provides the same features, with minimal change, while being
actively supported by a larger community.