fix: rename DeleteOutput to DisableOutput in computetask smart contract (#15)
At the time I created the disable output feature it was not tested as no
client was using it. Today I introduced an end to end test that use this
feature (https://github.com/Substra/substra-tests/pull/201) and it broke
in distributed mode because the smart contract entry `DisableOutput` did
not exit. This PR addresses that, by renaming the `DeleteOutput` smart
contract entry to `DisableOutput` which is the right name.

Signed-off-by: Alexandre-P <alexandre.picosson@owkin.com>