feat: assign worker from NewComputeTask (#18)
This feature will allow to specify the worker on task creation. That's a
first step to remove task-specific data and, eventually, task
categories.
For tasks having a datamanager, the worker is inferred from the
manager's owner and an error is returned if input mismatch.
For tasks not having input data, such as our current aggregation tasks,
explicitly setting the worker is mandatory.

For now, aggregate tasks will fallback to task-specific data if the
worker is not set.
This means that there is some throwaway code here and there: it will
disappear once all consumers have transitioned to the new field.