feat: modify computetask failure report (#277)
## Companion PR

- https://github.com/Substra/substra-backend/pull/727
- https://github.com/Substra/substra-frontend/pull/240

## Description
Modify `FailureReport`:
- add field `asset_type` containing the kind of asset the failure report
connect to
- rename `compute_task_key` to `asset_key`, which is a [wire compatible
change](https://groups.google.com/g/protobuf/c/hX4Mj0P4N0w) (i.e. does
not need to be declared as a new field)

## How has this been tested?

As this is going to be merged on a branch that is going to be merged to
a POC branch, we use MNIST as a baseline of a working model. We will
deal with failing tests on the POC before merging on main.

The e2e tests are also broken due to an issue on producing dumps during
release, but passed locally.

## Checklist

- [x] [changelog](../CHANGELOG.md) was updated with notable changes
- [ ] documentation was updated

---------

Signed-off-by: Guilhem Barthes <guilhem.barthes@owkin.com>
Signed-off-by: Guilhem Barthés <guilhem.barthes@owkin.com>