feat: add function event (#263)
## Companion PR

- https://github.com/Substra/substra-backend/pull/714

## Description

Add function events, used now we decoupled the building of the function
with the execution of the compute task. For that it add a status field
on the `Function`. It also includes another PR (merged here), to have
functions build logs working again.

Fixes FL-1160

## How has this been tested?

As this is going to be merged on a branch that is going to be merged to
a POC branch, we use MNIST as a baseline of a working model. We will
deal with failing tests on the POC before merging on main.

## Checklist

- [x] [changelog](../CHANGELOG.md) was updated with notable changes
- [ ] documentation was updated

---------

Signed-off-by: Guilhem Barthes <guilhem.barthes@owkin.com>
Signed-off-by: Guilhem Barthés <guilhem.barthes@owkin.com>
Signed-off-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: guilhem-barthes <guilhem-barthes@users.noreply.github.com>